refactor(core): extract library, log, and file stat operations from mobile stores to core
