+++ apps/mobile/src/components/FileCarousel/index.tsx
@@ -1,3 +1,4 @@
+import type { Category, SortBy, SortDir } from '@siastorage/core/db/operations'
 import { logger } from '@siastorage/logger'
 import * as ScreenOrientation from 'expo-screen-orientation'
 import { EyeIcon, EyeOffIcon } from 'lucide-react-native'
@@ -19,7 +20,6 @@ import { useFileStatus } from '../../lib/file'
 import { useToast } from '../../lib/toastContext'
 import { useFileCarousel } from '../../stores/fileCarousel'
 import type { FileRecord } from '../../stores/files'
-import type { Category, SortBy, SortDir } from '../../stores/library'
 import { toggleFavorite, useIsFavorite } from '../../stores/tags'
 import { palette } from '../../styles/colors'
 import BlocksLoader from '../BlocksLoader'
