+++ apps/mobile/src/stores/fileCarousel.test.ts
@@ -1,18 +1,11 @@
 import { act, renderHook, waitFor } from '@testing-library/react-native'
 import { database, db, initializeDB, resetDb } from '../db'
-import {
-  fetchFilePosition,
-  fetchFilesByIDs,
-  fetchSortedFileIds,
-  fetchTotalCount,
-  useFileCarousel,
-} from './fileCarousel'
+import { fetchFilesByIDs, useFileCarousel } from './fileCarousel'
 import { createFileRecord } from './files'
-import type { Category } from './library'
 
 import { invalidateCacheLibraryLists } from './librarySwr'
 
-describe('fileCarousel virtual list functions', () => {
+describe('fetchFilesByIDs', () => {
   const base = 1_000
 
   beforeEach(async () => {
@@ -26,503 +19,46 @@ describe('fileCarousel virtual list functions', () => {
 
   async function createRecord(params: {
     id: string
-    name: string | null
+    name: string
     createdAt: number
-    type?: string
-    addedAt?: number
-    size?: number
   }) {
     await createFileRecord({
       id: params.id,
-      name: params.name ?? `${params.id}.jpg`,
-      type: params.type ?? 'image/jpeg',
+      name: params.name,
+      type: 'image/jpeg',
       kind: 'file',
-      size: params.size ?? 100,
+      size: 100,
       hash: `hash-${params.id}`,
       createdAt: params.createdAt,
       updatedAt: params.createdAt,
       localId: null,
-      addedAt: params.addedAt ?? params.createdAt,
+      addedAt: params.createdAt,
       trashedAt: null,
       deletedAt: null,
     })
   }
 
-  async function seedDateRecords() {
+  test('returns files keyed by ID', async () => {
     await createRecord({ id: 'file-a', name: 'a.jpg', createdAt: base })
-    await createRecord({ id: 'file-b', name: 'b.jpg', createdAt: base + 10 })
     await createRecord({ id: 'file-c', name: 'c.jpg', createdAt: base + 20 })
-    await createRecord({ id: 'file-d', name: 'd.jpg', createdAt: base + 30 })
     await createRecord({ id: 'file-e', name: 'e.jpg', createdAt: base + 40 })
-  }
-
-  async function seedNameRecords() {
-    await createRecord({ id: 'id-3', name: 'a.jpg', createdAt: base })
-    await createRecord({ id: 'id-1', name: 'b.jpg', createdAt: base + 10 })
-    await createRecord({ id: 'id-5', name: 'c.jpg', createdAt: base + 20 })
-    await createRecord({ id: 'id-2', name: 'd.jpg', createdAt: base + 30 })
-    await createRecord({ id: 'id-4', name: 'e.jpg', createdAt: base + 40 })
-  }
-
-  async function seedAddedRecords() {
-    await createRecord({
-      id: 'file-a',
-      name: 'a.jpg',
-      createdAt: base,
-      addedAt: base + 40,
-    })
-    await createRecord({
-      id: 'file-b',
-      name: 'b.jpg',
-      createdAt: base + 10,
-      addedAt: base + 20,
-    })
-    await createRecord({
-      id: 'file-c',
-      name: 'c.jpg',
-      createdAt: base + 20,
-      addedAt: base + 30,
-    })
-    await createRecord({
-      id: 'file-d',
-      name: 'd.jpg',
-      createdAt: base + 30,
-      addedAt: base,
-    })
-    await createRecord({
-      id: 'file-e',
-      name: 'e.jpg',
-      createdAt: base + 40,
-      addedAt: base + 10,
-    })
-  }
-
-  async function seedSizeRecords() {
-    await createRecord({
-      id: 'file-a',
-      name: 'a.jpg',
-      createdAt: base,
-      size: 500,
-    })
-    await createRecord({
-      id: 'file-b',
-      name: 'b.jpg',
-      createdAt: base + 10,
-      size: 200,
-    })
-    await createRecord({
-      id: 'file-c',
-      name: 'c.jpg',
-      createdAt: base + 20,
-      size: 800,
-    })
-    await createRecord({
-      id: 'file-d',
-      name: 'd.jpg',
-      createdAt: base + 30,
-      size: 100,
-    })
-    await createRecord({
-      id: 'file-e',
-      name: 'e.jpg',
-      createdAt: base + 40,
-      size: 1000,
-    })
-  }
-
-  const defaultParams = {
-    sortBy: 'DATE' as const,
-    sortDir: 'DESC' as const,
-    categories: [] as Category[],
-    searchQuery: '',
-  }
-
-  describe('fetchTotalCount', () => {
-    test('returns correct count with no filters', async () => {
-      await seedDateRecords()
-      const count = await fetchTotalCount(defaultParams)
-      expect(count).toBe(5)
-    })
-
-    test('returns 0 for empty database', async () => {
-      const count = await fetchTotalCount(defaultParams)
-      expect(count).toBe(0)
-    })
-
-    test('filters by category', async () => {
-      await createRecord({
-        id: 'img-1',
-        name: 'photo1.jpg',
-        createdAt: base,
-        type: 'image/jpeg',
-      })
-      await createRecord({
-        id: 'vid-1',
-        name: 'video1.mp4',
-        createdAt: base + 10,
-        type: 'video/mp4',
-      })
-      await createRecord({
-        id: 'img-2',
-        name: 'photo2.jpg',
-        createdAt: base + 20,
-        type: 'image/jpeg',
-      })
-
-      const imageCount = await fetchTotalCount({
-        ...defaultParams,
-        categories: ['Image'],
-      })
-      expect(imageCount).toBe(2)
-
-      const videoCount = await fetchTotalCount({
-        ...defaultParams,
-        categories: ['Video'],
-      })
-      expect(videoCount).toBe(1)
-    })
+    const files = await fetchFilesByIDs(['file-a', 'file-c', 'file-e'])
+    expect(files.size).toBe(3)
+    expect(files.get('file-a')?.name).toBe('a.jpg')
+    expect(files.get('file-c')?.name).toBe('c.jpg')
+    expect(files.get('file-e')?.name).toBe('e.jpg')
   })
 
-  describe('fetchFilePosition', () => {
-    describe('DATE sorting (DESC)', () => {
-      test('returns correct position for middle file', async () => {
-        await seedDateRecords()
-        // DESC order: file-e(0), file-d(1), file-c(2), file-b(3), file-a(4)
-        const position = await fetchFilePosition('file-c', defaultParams)
-        expect(position).toBe(2)
-      })
-
-      test('returns 0 for first file', async () => {
-        await seedDateRecords()
-        const position = await fetchFilePosition('file-e', defaultParams)
-        expect(position).toBe(0)
-      })
-
-      test('returns last position for last file', async () => {
-        await seedDateRecords()
-        const position = await fetchFilePosition('file-a', defaultParams)
-        expect(position).toBe(4)
-      })
-
-      test('returns 0 for nonexistent file', async () => {
-        await seedDateRecords()
-        const position = await fetchFilePosition('nonexistent', defaultParams)
-        expect(position).toBe(0)
-      })
-    })
-
-    describe('DATE sorting (ASC)', () => {
-      test('returns correct position in ascending order', async () => {
-        await seedDateRecords()
-        // ASC order: file-a(0), file-b(1), file-c(2), file-d(3), file-e(4)
-        const position = await fetchFilePosition('file-c', {
-          ...defaultParams,
-          sortDir: 'ASC',
-        })
-        expect(position).toBe(2)
-      })
-
-      test('returns 0 for first file in ASC', async () => {
-        await seedDateRecords()
-        const position = await fetchFilePosition('file-a', {
-          ...defaultParams,
-          sortDir: 'ASC',
-        })
-        expect(position).toBe(0)
-      })
-    })
-
-    describe('NAME sorting (ASC)', () => {
-      test('returns correct position in alphabetical order', async () => {
-        await seedNameRecords()
-        // NAME ASC order: a.jpg(id-3), b.jpg(id-1), c.jpg(id-5), d.jpg(id-2), e.jpg(id-4)
-        const position = await fetchFilePosition('id-5', {
-          ...defaultParams,
-          sortBy: 'NAME',
-          sortDir: 'ASC',
-        })
-        expect(position).toBe(2)
-      })
-
-      test('returns 0 for first alphabetical file', async () => {
-        await seedNameRecords()
-        const position = await fetchFilePosition('id-3', {
-          ...defaultParams,
-          sortBy: 'NAME',
-          sortDir: 'ASC',
-        })
-        expect(position).toBe(0)
-      })
-    })
-
-    describe('NAME sorting (DESC)', () => {
-      test('returns correct position in reverse alphabetical order', async () => {
-        await seedNameRecords()
-        // NAME DESC order: e.jpg(id-4), d.jpg(id-2), c.jpg(id-5), b.jpg(id-1), a.jpg(id-3)
-        const position = await fetchFilePosition('id-5', {
-          ...defaultParams,
-          sortBy: 'NAME',
-          sortDir: 'DESC',
-        })
-        expect(position).toBe(2)
-      })
-    })
-
-    describe('ADDED sorting (DESC)', () => {
-      const addedParams = {
-        ...defaultParams,
-        sortBy: 'ADDED' as const,
-        sortDir: 'DESC' as const,
-      }
-
-      test('returns correct position for middle file', async () => {
-        await seedAddedRecords()
-        // addedAt values: a=base+40, c=base+30, b=base+20, e=base+10, d=base
-        // DESC order: file-a(0), file-c(1), file-b(2), file-e(3), file-d(4)
-        const position = await fetchFilePosition('file-b', addedParams)
-        expect(position).toBe(2)
-      })
-
-      test('returns 0 for first file', async () => {
-        await seedAddedRecords()
-        const position = await fetchFilePosition('file-a', addedParams)
-        expect(position).toBe(0)
-      })
-
-      test('returns last position for last file', async () => {
-        await seedAddedRecords()
-        const position = await fetchFilePosition('file-d', addedParams)
-        expect(position).toBe(4)
-      })
-    })
-
-    describe('SIZE sorting (DESC)', () => {
-      const sizeParams = {
-        ...defaultParams,
-        sortBy: 'SIZE' as const,
-        sortDir: 'DESC' as const,
-      }
-
-      test('returns correct position for middle file', async () => {
-        await seedSizeRecords()
-        // size values: e=1000, c=800, a=500, b=200, d=100
-        // DESC order: file-e(0), file-c(1), file-a(2), file-b(3), file-d(4)
-        const position = await fetchFilePosition('file-a', sizeParams)
-        expect(position).toBe(2)
-      })
-
-      test('returns 0 for first file', async () => {
-        await seedSizeRecords()
-        const position = await fetchFilePosition('file-e', sizeParams)
-        expect(position).toBe(0)
-      })
-
-      test('returns last position for last file', async () => {
-        await seedSizeRecords()
-        const position = await fetchFilePosition('file-d', sizeParams)
-        expect(position).toBe(4)
-      })
-    })
-
-    describe('tie-breaking by ID', () => {
-      test('breaks DATE ties using ID', async () => {
-        await createRecord({ id: 'file-c', name: 'c.jpg', createdAt: base })
-        await createRecord({ id: 'file-a', name: 'a.jpg', createdAt: base })
-        await createRecord({ id: 'file-b', name: 'b.jpg', createdAt: base })
-
-        // DESC with same date, order by ID DESC: file-c(0), file-b(1), file-a(2)
-        const positionC = await fetchFilePosition('file-c', defaultParams)
-        const positionB = await fetchFilePosition('file-b', defaultParams)
-        const positionA = await fetchFilePosition('file-a', defaultParams)
-
-        expect(positionC).toBe(0)
-        expect(positionB).toBe(1)
-        expect(positionA).toBe(2)
-      })
-
-      test('breaks NAME ties using ID', async () => {
-        await createRecord({ id: 'file-c', name: 'same.jpg', createdAt: base })
-        await createRecord({
-          id: 'file-a',
-          name: 'same.jpg',
-          createdAt: base + 10,
-        })
-        await createRecord({
-          id: 'file-b',
-          name: 'same.jpg',
-          createdAt: base + 20,
-        })
-
-        // NAME ASC with same name, order by ID ASC: file-a(0), file-b(1), file-c(2)
-        const params = {
-          ...defaultParams,
-          sortBy: 'NAME' as const,
-          sortDir: 'ASC' as const,
-        }
-        const positionA = await fetchFilePosition('file-a', params)
-        const positionB = await fetchFilePosition('file-b', params)
-        const positionC = await fetchFilePosition('file-c', params)
-
-        expect(positionA).toBe(0)
-        expect(positionB).toBe(1)
-        expect(positionC).toBe(2)
-      })
-    })
-
-    describe('category filtering', () => {
-      test('returns correct position with category filter', async () => {
-        await createRecord({
-          id: 'img-1',
-          name: 'photo1.jpg',
-          createdAt: base,
-          type: 'image/jpeg',
-        })
-        await createRecord({
-          id: 'vid-1',
-          name: 'video1.mp4',
-          createdAt: base + 10,
-          type: 'video/mp4',
-        })
-        await createRecord({
-          id: 'img-2',
-          name: 'photo2.jpg',
-          createdAt: base + 20,
-          type: 'image/jpeg',
-        })
-        await createRecord({
-          id: 'vid-2',
-          name: 'video2.mp4',
-          createdAt: base + 30,
-          type: 'video/mp4',
-        })
-        await createRecord({
-          id: 'img-3',
-          name: 'photo3.jpg',
-          createdAt: base + 40,
-          type: 'image/jpeg',
-        })
-
-        // Images only DESC: img-3(0), img-2(1), img-1(2)
-        const position = await fetchFilePosition('img-2', {
-          ...defaultParams,
-          categories: ['Image'],
-        })
-        expect(position).toBe(1)
-      })
-
-      test('returns 0 when file filtered out by category', async () => {
-        await createRecord({
-          id: 'img-1',
-          name: 'photo.jpg',
-          createdAt: base,
-          type: 'image/jpeg',
-        })
-
-        const position = await fetchFilePosition('img-1', {
-          ...defaultParams,
-          categories: ['Video'],
-        })
-        expect(position).toBe(0)
-      })
-    })
+  test('returns empty map for empty ID list', async () => {
+    const files = await fetchFilesByIDs([])
+    expect(files.size).toBe(0)
   })
 
-  describe('fetchSortedFileIds', () => {
-    test('returns IDs in sort order', async () => {
-      await seedDateRecords()
-      // DESC order: file-e, file-d, file-c, file-b, file-a
-      const ids = await fetchSortedFileIds(defaultParams, 5, 0)
-      expect(ids).toEqual(['file-e', 'file-d', 'file-c', 'file-b', 'file-a'])
-    })
-
-    test('respects limit and offset', async () => {
-      await seedDateRecords()
-      const ids = await fetchSortedFileIds(defaultParams, 2, 1)
-      expect(ids).toEqual(['file-d', 'file-c'])
-    })
-
-    test('returns empty array for empty database', async () => {
-      const ids = await fetchSortedFileIds(defaultParams, 10, 0)
-      expect(ids).toEqual([])
-    })
-
-    test('returns IDs in ADDED sort order', async () => {
-      await seedAddedRecords()
-      // addedAt values: a=base+40, c=base+30, b=base+20, e=base+10, d=base
-      const ids = await fetchSortedFileIds(
-        { ...defaultParams, sortBy: 'ADDED', sortDir: 'DESC' },
-        5,
-        0,
-      )
-      expect(ids).toEqual(['file-a', 'file-c', 'file-b', 'file-e', 'file-d'])
-    })
-
-    test('returns IDs in SIZE sort order', async () => {
-      await seedSizeRecords()
-      // size values: e=1000, c=800, a=500, b=200, d=100
-      const ids = await fetchSortedFileIds(
-        { ...defaultParams, sortBy: 'SIZE', sortDir: 'DESC' },
-        5,
-        0,
-      )
-      expect(ids).toEqual(['file-e', 'file-c', 'file-a', 'file-b', 'file-d'])
-    })
-  })
-
-  describe('fetchFilesByIDs', () => {
-    test('returns files keyed by ID', async () => {
-      await seedDateRecords()
-      const files = await fetchFilesByIDs(['file-a', 'file-c', 'file-e'])
-      expect(files.size).toBe(3)
-      expect(files.get('file-a')?.name).toBe('a.jpg')
-      expect(files.get('file-c')?.name).toBe('c.jpg')
-      expect(files.get('file-e')?.name).toBe('e.jpg')
-    })
-
-    test('returns empty map for empty ID list', async () => {
-      const files = await fetchFilesByIDs([])
-      expect(files.size).toBe(0)
-    })
-
-    test('skips nonexistent IDs', async () => {
-      await seedDateRecords()
-      const files = await fetchFilesByIDs(['file-a', 'nonexistent'])
-      expect(files.size).toBe(1)
-      expect(files.get('file-a')?.id).toBe('file-a')
-    })
-  })
-
-  describe('edge cases', () => {
-    test('handles single file', async () => {
-      await createRecord({ id: 'only-file', name: 'solo.jpg', createdAt: base })
-
-      const count = await fetchTotalCount(defaultParams)
-      const position = await fetchFilePosition('only-file', defaultParams)
-      const files = await fetchFilesByIDs(['only-file'])
-
-      expect(count).toBe(1)
-      expect(position).toBe(0)
-      expect(files.get('only-file')?.id).toBe('only-file')
-    })
-
-    test('handles files with null names', async () => {
-      await createRecord({ id: 'file-null', name: null, createdAt: base })
-      await createRecord({
-        id: 'file-named',
-        name: 'named.jpg',
-        createdAt: base + 10,
-      })
-
-      const count = await fetchTotalCount({
-        ...defaultParams,
-        sortBy: 'NAME',
-        sortDir: 'ASC',
-      })
-      expect(count).toBe(2)
-
-      const files = await fetchFilesByIDs(['file-null', 'file-named'])
-      expect(files.size).toBe(2)
-    })
+  test('skips nonexistent IDs', async () => {
+    await createRecord({ id: 'file-a', name: 'a.jpg', createdAt: base })
+    const files = await fetchFilesByIDs(['file-a', 'nonexistent'])
+    expect(files.size).toBe(1)
+    expect(files.get('file-a')?.id).toBe('file-a')
   })
 })
