+++ apps/mobile/src/stores/fileStats.ts
@@ -1,15 +1,7 @@
+import type { UploadCategoryStats } from '@siastorage/core/db/operations'
+import * as ops from '@siastorage/core/db/operations'
 import { db } from '../db'
 import { getIndexerURL } from './settings'
-
-export type UploadCategoryStats = {
-  total: number
-  remaining: number
-  uploaded: number
-  totalBytes: number
-  uploadedBytes: number
-  percent: string
-  percentDecimal: number
-}
 
 export type UploadStats = {
   overall: UploadCategoryStats
@@ -20,56 +12,6 @@ export type UploadStats = {
   docs: UploadCategoryStats
   other: UploadCategoryStats
   thumbnails: UploadCategoryStats
-}
-
-async function counts(
-  where: string,
-  indexerURL: string,
-): Promise<{
-  total: number
-  remaining: number
-  uploaded: number
-  totalBytes: number
-  uploadedBytes: number
-  percent: string
-  percentDecimal: number
-}> {
-  const totalRow = await db().getFirstAsync<{
-    count: number
-    totalBytes: number
-  }>(
-    `SELECT COUNT(*) as count, COALESCE(SUM(f.size), 0) as totalBytes FROM files f WHERE ${where}`,
-  )
-  const remainingRow = await db().getFirstAsync<{
-    count: number
-    totalBytes: number
-  }>(
-    `SELECT COUNT(*) as count, COALESCE(SUM(f.size), 0) as totalBytes
-     FROM files f
-     WHERE ${where}
-       AND NOT EXISTS (
-         SELECT 1 FROM objects o
-         WHERE o.fileId = f.id AND o.indexerURL = ?
-       )`,
-    indexerURL,
-  )
-  const total = totalRow?.count ?? 0
-  const totalBytes = totalRow?.totalBytes ?? 0
-  const remaining = remainingRow?.count ?? 0
-  const remainingBytes = remainingRow?.totalBytes ?? 0
-  const uploaded = Math.max(0, total - remaining)
-  const uploadedBytes = Math.max(0, totalBytes - remainingBytes)
-  const percentDecimal = totalBytes ? uploadedBytes / totalBytes : 1
-  const percent = `${(percentDecimal * 100).toFixed(1)}%`.padStart(6)
-  return {
-    total,
-    remaining,
-    uploaded,
-    totalBytes,
-    uploadedBytes,
-    percent,
-    percentDecimal,
-  }
 }
 
 /** Get upload stats for the library with category breakdown. */
@@ -83,13 +25,15 @@ export async function getUploadStats(): Promise<UploadStats> {
   const otherWhere = `f.kind = 'file' AND ${active} AND f.type NOT LIKE 'image/%' AND f.type NOT LIKE 'video/%' AND f.type NOT LIKE 'audio/%' AND f.type NOT LIKE 'text/%' AND f.type NOT LIKE 'application/%'`
   const thumbsWhere = `f.kind = 'thumb' AND ${active}`
 
+  const q = (where: string) =>
+    ops.queryUploadCategoryStats(db(), where, indexerURL)
   const [photosC, videosC, audioC, docsC, otherC, thumbsC] = await Promise.all([
-    counts(photosWhere, indexerURL),
-    counts(videosWhere, indexerURL),
-    counts(audioWhere, indexerURL),
-    counts(docsWhere, indexerURL),
-    counts(otherWhere, indexerURL),
-    counts(thumbsWhere, indexerURL),
+    q(photosWhere),
+    q(videosWhere),
+    q(audioWhere),
+    q(docsWhere),
+    q(otherWhere),
+    q(thumbsWhere),
   ])
 
   const fileCategories = [photosC, videosC, audioC, docsC, otherC]
